Start KDE integration by "borrowing" oxygen icons for those missing :-)
They're currently copied locally, I guess when we do a proper KDE integration
they'll have to be gathered from the system, if we are on a system
where KDE is installed. For that reason I kept the original names. I think it would make
sense to replace as many icons as possible with oxygen icons. They look good and it will 
help later to integrate into KDE.
Some old icons are lying around, does anybody mind if I clean them up?


git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@1177319 283d02a7-25f6-0310-bc7c-ecb5cbfe19da