A new feature:  annotation view items can now be set to have their positions 'fixed to data', so that if you scroll or more data comes in, the annotation object moves along with the data.  It is set in the dimensions tab of the view item's dialog.
What I am commiting here works ~perfectly for boxes, circles and
ellipses for both scaling and translation.

It works fine for labels for translation, but they tend to drift with
scaling.

For arrows and lines, it works fine for translation, but has problems
with scaling, because, for view objects, the item's rotation is defined
in screen units.  This will be fixed.


git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@1279327 283d02a7-25f6-0310-bc7c-ecb5cbfe19da