Add drag&drop feature: - if dropping a .kst file, open it after checking what needed to be saved has been saved - if dropping something else, open it in the datawizard
The prompt to save feature in case of change has been refactored along the way to make it more readable code-wise. I hope I haven't broken anything.
The reason why our previous attempt had not been conclusive is that we have to setAcceptDrops(false) all the way down the parent chain: _tabWidget *and* View.

Note: my tests exposed a bug: adding view items is not registered as a change. This should be fixed!

FEATURE: 272510


git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@1284152 283d02a7-25f6-0310-bc7c-ecb5cbfe19da