Make sure that adding view items is registered as a change.
Barth, please review: I'm not familiar with the undo/redo code and I'm not sure I put it in the right place. In particular, when adding an item, hitting undo and 
trying to open a file it prompts for saving while it actually shouldn't. 
I'm committing this anyway as the current behavior is better than not prompting for save at all when view item(s) have been added. But we can possibly make it even 
better.


git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@1284792 283d02a7-25f6-0310-bc7c-ecb5cbfe19da