Implement automatic "unpacking" of netCDF vectors of type ncShort if they have scale_factor and add_offset attributes, as per the convention described in: http://www.unidata.ucar.edu/software/netcdf/docs/netcdf/Attribute-Conventions.html
Leveraging all the metadata already implemented, this approach is much
more efficient and takes much less code than the one originally
proposed. And it seems to work very well.

Thanks to Patrick R. for the idea, and the code even though in the end
it was easier to rewrite it differently.

I still have to think about the other part of the patch (matrix
components as vectors)...


git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@1355983 283d02a7-25f6-0310-bc7c-ecb5cbfe19da