Fix data source config changes.
If a data source configuration is changed, apply the change
to all dependent vectors.

note: it used to be that editing a data source produced a new
data source - apparently so that you could have vectors reading
from the same file with different data source configurations.
This seems confusing and didn't work quite right.