Rewrite a bunch of scripting...
The old script interface relied on dialogs, which meant
that scripting would break if used with translations.

The new script interface does not have this problem.
This commit deletes most of the old system (though some
dangling stuff still remains for now) and re-implements
view object and primitives.  DataObjects and Curves/Images
are not yet implementd.  Nor are editable matrixes and
vectors.