This is a new plugin designed to read FITS tables.  Note, to compile KST on my machine (running OS X 10.6.8) required me to remove the KST_THREAD_LOCAL keywords from kst_atof.h and kst_atof.cpp.  This was because my compiler does not seem to support "thread local storage".
Secondly, I had to disable the fitsimage plugin in
src/datasources/CMakeLists.txt.  For some reason, I got errors as if the
cfitsio libraries weren't being linked properly.  This is most puzzling since
I used the cfitsio libraries in fitstable, and they link just fine.

Because I had to disable fitsimage, I can't test how fitsimage and fitstable
interact when loading a file with the data wizard.  Will fitsimage or
fitstable get control of the file?  fitstable will read the file, and if
it is a fits file with at least one table extension with more than zero
rows, then it will return a value of 80, indicating it can read the file.