Use "native" if "use raster" is not selected.
QT's default graphics system on all platforms has changed to
"raster" so the code to select "raster" or not did nothing.
However, native works far better for remote X, so it is
a very good idea to have it available.  This patch makes
sure that the selection box actually does something.