Correctly read old kst sessions.
Bump the kst file id to 2.0.9 and fix loading 2.0 sessions.

Kst sessions pack slave scalar names in such a way that new scalars can
not be added to vectors once the vector is created and the scalars are
named.

In an earlier commit, new scalars (iMin and iMax) were added to vectors.
This breaks loading of old sessions into new kst.

With this patch, if kst is loading an old session version, it will
not add iMin and iMax to the vectors.  It will only add them for
new sessions.

Note: once a session has been created as a 2.0 session, it can never
be bumped to a 2.0.9 or later session.

This kind of sucks, but at least old sessions still work as old
sessions, and new one work and new ones.

This may be a temporary fix...