Add image streams to data sources.
Data sources now support 'image streams' (imagine video streams).

Also add the BIS data source (BIT Image Stream).
In version 0, this is a highly custom data source to be used
by the BIT telescope. It is mainly only useful here as an
example of how to write a data source that can handle image
streams.  It is super light-weight.