Improve UI with a warning of inconsistent X vector.
BUG: 356255
Re-using an x vector from a previous plot is probably
what you want to do if you have kept the data range
the same, but probably not what you wanted if you
have changed the data range.  Add a warning and a
suggestion on how to fix it if your selected x vector
does not match the new Y vectors in the datawizard.