NaNs no longer pollute fits and filters.
Formally, if you have NaNs in your data, then the results
of a fit or a fourier based filter should be all NaNs as well.

However, it is very unlikely that this is what you actually
wanted.  This patch automatically interpolates over NaNs in
many operations that would otherwise be polluted.

This is a step on the path to allowing for filter vectors.