Fix serious performance issue in some cases.
On some systems and on some versions of Qt5, QDrawPolyLine() is
extremely slow for large windows (4k).  This replaces DrawPolyLine()
with a bunch of DrawLines, which dramatically improves performance in
these cases.  However, it also means that wide lines with transparency
will have artifacts, and dashed lines will not work right when the point
density is very high.  So this is not a permanent fix.  The issue has
been reported to Qt.