Fix negative-going generated vectors in .kst files and vector dialog.
Negative-going (monotonically decreasing) generated vectors are
implemented by not assuming min==first and max==last.  This involves a
change to the .kst file by adding attributes "first" and "last" to the
<generatedvector> tag.  The "min" and "max" attributes are kept in the
tag for backwards compatibility.

A negative-going generated vector from 1 to 0 will be saved like:

      <generatedvector first="1" last="0" min="0" max="1" [...]

Previously this would have been just:

      <generatedvector min="0" max="1" [...]

This is sort-of backwards compatibile in the sense that an old kst will
simply ignore the first and last attributes and create a increasing vector
from 0 to 1, which is only sort-of right, becasue the created vector will
be backwards, but at least kst isn't going to reject the .kst file.

This commit changes the XML parsing so that:

  - if "first" exists, then the generated vector will use "first" and
    "last" as the start and end points of the vector.  In this case,
    the "min" and "max" attributes are ignored.

  - otherwise, it will use "min" as the start and "max" as the end
    point of the vector, creating a non-negative-going vector, as
    before.